pyproject.toml
[project]
name = "aichatbot-backend"
version = "0.1.0"
description = "Modular FastAPI backend for multi-provider chat"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.32",
  "asyncpg>=0.30.0",
  "pydantic-settings>=2.4.0",
  "PyJWT>=2.9.0",
  "python-json-logger>=2.0.7",
  "tenacity>=9.0.0",
  "openai>=1.43.0",
  "anthropic>=0.42.0",
  "httpx>=0.27.0",
  "email-validator>=2.2.0",
  "prometheus-client>=0.20.0",
  "greenlet>=3.1.1",
]

[dependency-groups]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "respx>=0.21.1",
  "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
